+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Tests for "War".
+- Tests for "Dwarfs standing on the shoulders of giants".
 
 ## [2.2.0] - 2022-02-10
 ### Added